requirements_test.txt
requests>=2.28.1
voluptuous>=0.13.1
homeassistant>=2023.1.0
pycryptodome>=3.18.0
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0